requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==3.7.1
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comtypes==1.4.11; sys_platform == "win32"
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.104.1
Flask==3.1.1
flask-cors==6.0.1
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
MouseInfo==0.1.3
multidict==6.6.4
openai==0.28.1
pillow==11.3.0
propcache==0.3.2
pvporcupine==3.0.5
PyAudio==0.2.14
PyAutoGUI==0.9.54
pydantic==2.11.7
pydantic_core==2.33.2
PyGetWindow==0.0.9
pyjokes==0.8.3
PyMsgBox==1.0.9
pyperclip==1.9.0
pypiwin32==223; sys_platform == "win32"
PyRect==0.2.0
PyScreeze==1.0.1
python-dotenv==1.0.0
python-multipart==0.0.20
pyttsx3==2.99
pytweening==1.2.0
pywhatkit==5.4
pywin32>=306; sys_platform == "win32"
PyYAML==6.0.2
requests==2.32.4
sniffio==1.3.1
soupsieve==2.7
SpeechRecognition==3.14.3
starlette==0.27.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.24.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
wikipedia==1.4.0
yarl==1.20.1
